+++ apps/web/lib/ai/generate-filters.ts
